Move NoteItem to molecuels
