+++ src/components/Tutorials/TutorialsNoteItem.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Link from '../atoms/Link'
 import cc from 'classcat'
 import { TutorialsNavigatorTreeItem } from '../../lib/tutorials'
-import { StyledNoteListItem } from '../organisms/NoteList/NoteItem'
+import { StyledNoteListItem } from '../molecules/NoteItem'
 
 type TutorialsNoteItemProps = {
   note: TutorialsNavigatorTreeItem
