--- src/components/organisms/NoteList/index.ts
@@ -1,3 +1,0 @@
-import NoteList from './NoteList'
-
-export default NoteList
